Extract shared cloze parser used by ExamplePicker and hint mapper
Both ExamplePicker.clozeText() and the new translation hint mapper
parsed <w>-tagged text into (plainText, answerRanges) the same way.
Pull the logic out into parseClozeFromTaggedText() so there is a
single implementation, and switch the mapper helper from an extension
to a regular function.
